docs: Update README with MySQL management and migration steps
Added detailed instructions to the README.md for managing MySQL on the hosting company's server, including steps to run MySQL in a Docker container locally, open a shell in the container, create a dump file, connect to the MySQL server, copy the dump file out of the container, and restore/import the database on a new server. Also documented a modification to the /wp/wp-config.php file to point to the new MySQL 8.x database as of 08/24/2024.