Update README with EF Core setup and migration steps
README.md has been updated with new instructions as of 10/13/2024. These instructions cover the use of EF Core Power Tools for reverse engineering the Odyssey Registration database. Additionally, a link to seed data for the `School` entity has been added to the `OnModelCreating` method in the `OdysseyContext` class. EF Core migrations were added using Developer PowerShell for Visual Studio 2022, with specific commands for adding and updating migrations and the database.